+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add possibility to run cert-operator as a unique app, reconciling special version '0.0.0'.
 
+### Fixed
+
+- Avoid including certconfig UID in organizations for kubeconfig requests.
+
 ## [2.0.1] - 2022-04-04
 
 ### Fixed